requirements.txt
numpy
pandas
torch
transformers==4.34.1
accelerate==0.24.0
evaluate
datasets
packaging
sentencepiece
scikit-learn
wandb
mosaicml-streaming==0.7.2
awscli
ninja
wheel
rouge-score
PyYAML
tqdm
flash-attn==2.3.3
